+++ src/main/java/cat/soft/src/parking/controller/room/RoomService.java
@@ -44,13 +44,13 @@ public class RoomService {
 	public PostCreateRoomRes createRoom(String token) {
 		User user = userRepository.findUsersByEmail(jwtTokenProvider.getEmail(token));
 		if (user == null) {
-			return new PostCreateRoomRes(0);
+			return new PostCreateRoomRes(0L);
 		}
 		if (user.getRoomIdx() != 0) {
-			return new PostCreateRoomRes(0);
+			return new PostCreateRoomRes(0L);
 		}
 		if (user.getRole() != 0) {
-			return new PostCreateRoomRes(0);
+			return new PostCreateRoomRes(0L);
 		}
 		Room room = roomRepository.save(Room.builder().idx(user.getIdx()).build());
 		user.setRoomIdx(room.getIdx());
@@ -60,23 +60,23 @@ public class RoomService {
 		return new PostCreateRoomRes(user.getIdx());
 	}
 
-	public GetJoinRoomRes joinRoom(Integer roomId, String token) {
+	public GetJoinRoomRes joinRoom(Long roomId, String token) {
 		User user = userRepository.findUsersByEmail(jwtTokenProvider.getEmail(token));
 		if (user == null) {
 			return null; // 5000 불가능한 유저
 		}
 		if (user.getRole() != 0) {
-			return new GetJoinRoomRes(0); // [5001] 승인된 유저입니다.
+			return new GetJoinRoomRes(0L); // [5001] 승인된 유저입니다.
 		}
 		if (user.getRoomIdx() != 0) {
-			return new GetJoinRoomRes(-1); // [5002] 승인 대기중 입니다.
+			return new GetJoinRoomRes(-1L); // [5002] 승인 대기중 입니다.
 		}
 		if (roomId == null) {
-			return new GetJoinRoomRes(-2); // [5003] 존재하지 않는 방 입니다.
+			return new GetJoinRoomRes(-2L); // [5003] 존재하지 않는 방 입니다.
 		}
 		Room room = roomRepository.findById(roomId).orElse(null);
 		if (room == null) {
-			return new GetJoinRoomRes(-3); // [5003] 존재하지 않는 방 입니다.
+			return new GetJoinRoomRes(-3L); // [5003] 존재하지 않는 방 입니다.
 		}
 		user.setRoomIdx(roomId);
 		userRepository.save(user);
@@ -94,7 +94,7 @@ public class RoomService {
 		return new GetQrCheckRes(user.getRoomIdx());
 	}
 
-	public GetUserListByAdminRes userListByAdmin(Integer roomId, String token) {
+	public GetUserListByAdminRes userListByAdmin(Long roomId, String token) {
 		User admin = userRepository.findUsersByEmail(jwtTokenProvider.getEmail(token));
 		Room room = roomRepository.findById(roomId).orElse(null);
 		if (admin == null || room == null) {
@@ -125,7 +125,7 @@ public class RoomService {
 		return new GetUserListByAdminRes(newUserInfo, oldUserInfo);
 	}
 
-	public PutUserApproveRes approveUser(Integer roomId, PutUserApproveReq req, String token) {
+	public PutUserApproveRes approveUser(Long roomId, PutUserApproveReq req, String token) {
 		Room room = roomRepository.findById(roomId).orElse(null);
 		User admin = userRepository.findUsersByEmail(jwtTokenProvider.getEmail(token));
 		User user = userRepository.findById(req.getUserIdx()).orElse(null);
@@ -137,18 +137,18 @@ public class RoomService {
 			return null;
 		}
 		if (Objects.equals(user.getIdx(), admin.getIdx())) { // 셀프 추방 금지
-			return new PutUserApproveRes(-1);
+			return new PutUserApproveRes(-1L);
 		}
 		if (!Objects.equals(user.getRoomIdx(), admin.getRoomIdx())) { // 신청하지 않은 유저 변경 금지
-			return new PutUserApproveRes(-2);
+			return new PutUserApproveRes(-2L);
 		}
 		if (req.getRole() == 0) { // 거절
 			user.setRole(Long.valueOf(req.getRole()));
-			user.setRoomIdx(0);
+			user.setRoomIdx(0L);
 		} else if (req.getRole() == 1) { // 승인
 			user.setRole(Long.valueOf(req.getRole()));
 		} else {
-			return new PutUserApproveRes(-3);
+			return new PutUserApproveRes(-3L);
 		}
 		userRepository.save(user);
 		return new PutUserApproveRes(user.getIdx());
