+++ src/main/java/cat/soft/src/parking/controller/user/UserService.java
@@ -31,14 +31,14 @@ public class UserService {
 	@Autowired
 	private JwtTokenProvider jwtTokenProvider;
 
-	public PutUserInfoRes updateUserInfo(@PathVariable Integer id, @RequestBody PutUserInfoReq userInfoReq,
+	public PutUserInfoRes updateUserInfo(@PathVariable Long id, @RequestBody PutUserInfoReq userInfoReq,
 		String token) {
 		User user = userRepository.findUsersByEmail(jwtTokenProvider.getEmail(token));
 		if (user == null) {
 			return null;
 		}
 		if (!Objects.equals(id, user.getIdx()))
-			return new PutUserInfoRes(0);
+			return new PutUserInfoRes(0L);
 		UserInfo userInfo = userInfoRepository.findById(user.getIdx()).orElse(null);
 		if (userInfo == null) {
 			return null;
@@ -55,7 +55,7 @@ public class UserService {
 		return new PutUserInfoRes(user.getIdx());
 	}
 
-	public GetUserInfoRes getUserInfo(Integer id, String token) {
+	public GetUserInfoRes getUserInfo(Long id, String token) {
 		User user = userRepository.findUsersByEmail(jwtTokenProvider.getEmail(token));
 		if (user == null) {
 			return null;
