+++ src/main/java/cat/soft/src/parking/model/Report.java
@@ -22,19 +22,19 @@ public class Report {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "idx", nullable = false)
-	private Integer idx;
+	private Long idx;
 
 	@Column(name = "victim", nullable = false)
-	private Integer victim;
+	private Long victim;
 
 	@Column(name = "suspect", nullable = false)
-	private Integer suspect;
+	private Long suspect;
 
 	@Column(name = "time", nullable = false)
 	private ZonedDateTime time;
 
 	@Builder
-	public Report(Integer victim, Integer suspect, ZonedDateTime time) {
+	public Report(Long victim, Long suspect, ZonedDateTime time) {
 		this.victim = victim;
 		this.suspect = suspect;
 		this.time = time;
